improve machine error when running on Linux
On Linux, Podman Desktop can talk to the Podman socket of the host and
to the ones of the Linux VM.  As outlinted in #42, talking to socket of
the host's Podman is currently not supported by the extension.  While
logging into the container registry is perfectly fine, running
subscriptio-manager is not; it requires root privileges to install and
run it.  So instead of a silent error, throw a user-visible error
indicating that it's currently not supported on a Linux host.

Fixes: #42
Signed-off-by: Valentin Rothberg <vrothberg@redhat.com>